+++ nonebot/drivers/fastapi.py
@@ -21,13 +21,12 @@ from starlette.websockets import WebSocket, WebSocketState, WebSocketDisconnect
 
 from nonebot.config import Env
 from nonebot.typing import overrides
+from nonebot.internal.model import FileTypes
 from nonebot.exception import WebSocketClosed
 from nonebot.config import Config as NoneBotConfig
 from nonebot.drivers import Request as BaseRequest
 from nonebot.drivers import WebSocket as BaseWebSocket
 from nonebot.drivers import ReverseDriver, HTTPServerSetup, WebSocketServerSetup
-
-from ..internal.model import FileTypes
 
 
 def catch_closed(func):