+++ website/src/pages/changelog.md
@@ -15,7 +15,7 @@ description: Changelog
 - 新增 `load_builtin_plugin` 函数，用于加载单个内置插件
 - 修改 `Message` 和 `MessageSegment` 类，完善 typing，转移 Mapping 构建支持至 pydantic validate
 - 调整项目结构，分离内部定义与用户接口
-- 修改 `Bot Connection Hook` 支持依赖注入
+- 新增 Bot 连接事件钩子 (如 `driver.on_bot_connect` ) 的依赖注入
 
 ## v2.0.0b1